size_t for GetKmerIndex position

+++ ntcoding.cpp
@@ -58,7 +58,7 @@ void GenerateShapePos (std::string shape) {
     }
 }
 
-uint32_t GetKmerIndexAtPos (char* sequence, uint32_t pos, uint32_t seed_size) {
+uint32_t GetKmerIndexAtPos (char* sequence, size_t pos, uint32_t seed_size) {
     uint32_t kmer = 0;
 
     bool N_char = false;
